feat(client,core,server): support explicit null values, add camera capture, theming events, and rework file handling
  - data: allow explicit null writes/reads across UpdateData/GetData for in-memory
    data sources (Explicit.ExplicitValue now nullable, MosaicSerializer polymorphic
    registration completed, DataSourceSchema.Inline added); remove the unused/broken
    ExplicitNull sentinel in favor of plain Kotlin null end-to-end
  - camera: add TakePicture and GetImageFromGalery events with per-platform
    CameraManager (Android/iOS/JVM/WasmJs), CompressionScheme and
    ImageResizeOptions for output control
  - files: generalize PlatformFileHandler across platforms, add
    DeleteFile/GetFile/SaveFile use cases, FileBytesDecoding helper, and a new
    DownloadFileToDisk event; wire Android FileProvider (manifest + file paths)
  - theme: rework SetTheme/ResetTheme event schema, builder, holder and runner
  - network: fix chunked file upload/UploadPlatformFile handling across targets
  - tiles: add PopupTileSchema; tweak AsyncImage and DropdownList tile schemas
  - chore: drop unused OnDownloadPartialEventTrigger; update events/tiles/triggers
    catalogs and the mosaic skill docs to reflect the above

+++ mosaic-client/src/commonMain/kotlin/dev/catbit/mosaic/client/data/data_sources/network/UploadPlatformFile.kt
@@ -20,5 +20,5 @@ internal expect suspend fun uploadPlatformFile(
     platformFile: PlatformFile,
     contentType: String?,
     queryParameters: Map<String, Any?>?,
-    onProgress: suspend (Int) -> Unit
+    onProgress: suspend (Float) -> Unit
 ): UploadResult
